% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_team_strength_formula.R
\name{add_team_strength_formula}
\alias{add_team_strength_formula}
\title{Add Team Strength Formula}
\usage{
add_team_strength_formula(forumula_prefix, data)
}
\arguments{
\item{forumula_prefix}{Formula in String Form}

\item{data}{}
}
\value{
Formula with the Strength (Difference Form Indicators ) appended to the input formula.
}
\description{
Add Strength (Difference Form Indicators) to a Formula
}
\examples{
add_team_strength_formula("HomePointDifferential~Jamsil+PartialFans + NoFans + Playoffs", data)

}
